[Profile creation] Create Profile object in the controller
This CL is a prefactoring for making a LaCrOS implementation of the
profile creation flow. It moves the logic to create a profile used for
the dice flow into the controller. This encapsulates the dice-specific
logic. As the next step, the controller will be split into a dice
specific and a platform-independent part.

Bug: 1227029
Change-Id: I52de8a3d233fce11733d81a3772af0a03e5d6219
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3018204
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Jan Krcal <jkrcal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#908896}