Remove WebUsbPrerenderinBrowserTest.ShowChooserInPrerenderin
This test is flaky and after discussing with the Prerendering
team there isn't value in keeping it because triggering the
chooser UI is not possible for a page being pre-renderered to
show the chooser because it requires user activation. A test
which injected a user activation would be testing an unrealistic
scenario. We already have WPTs for calling getDevices() during
pre-rendering.

Bug: 1236542
Change-Id: Idd1c9be68d23f8fe4258af72fd321ca69f315443
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3071735
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Commit-Queue: Lucas Gadani <lfg@chromium.org>
Auto-Submit: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Lucas Gadani <lfg@chromium.org>
Reviewed-by: Ken Rockot <rockot@google.com>
Cr-Commit-Position: refs/heads/master@{#908937}