Move kEpsilon from DelegatedInkPointRendererBase to DelegatedInkPoint
|kEpsilon| was added to DelegatedInkPointRendererBase to allow for
some small precision errors causing DelegatedInkPoint and
DelegatedInkMetadata to have the same timestamp, but slightly different
points. However, this epsilon value is needed in the gl namespace as
well, because the OS API path is seeing a similar issue. Therefore, this
CL moves the epsilon to instead just live on DelegatedInkPoint so that
it can be used anywhere. This also added a convenient method in each
DelegatedInk* class so that Point and Metadata can be compared more
easily.

Bug: 1052145
Change-Id: I0ac07c67ebec0bb36f277724d8fadcb13773e6a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3060054
Commit-Queue: Mario Bianucci <mabian@microsoft.com>
Reviewed-by: Daniel Libby <dlibby@microsoft.com>
Reviewed-by: weiliangc <weiliangc@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Cr-Commit-Position: refs/heads/master@{#908985}