[CrOS Bluetooth] Add new bluetooth device detail page
Adds a new page <os-settings-bluetooth-device-detail-subpage> and a new
route BLUETOOTH_DEVICE_DETAIL to <os-settings-bluetooth-page>. This CL
does not add test for <os-settings-bluetooth-device-detail-subpage>,
test would be added in a subsequent CL, once more functionality is
added. The route is currently not being used, it will be used to
navigate to the device detail after
<os-settings-bluetooth-devices-subpage>, has been added.

Note: This route was tested by navigation in os-bluetooth-summary page
to os-settings-bluetooth-device-detail-subpage.

Bug: 1010321
Change-Id: Id0c0b544f41d3f212c75a802c118043c43e1e608
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3068166
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Gordon Seto <gordonseto@google.com>
Commit-Queue: Nnamdi Theodore Johnson-kanu <tjohnsonkanu@google.com>
Cr-Commit-Position: refs/heads/master@{#908986}