app_list: Auto-scroll when dragging apps to top/bottom of bubble
When the user drags an icon below the bottom of the ScrollView, if the
view can scroll in that direction, start scrolling down. Do the
opposite at the top of the Scroll.

Attempt to scroll once per frame while the mouse is in that margin.
Scroll at a constant rate in dips-per-second by correcting for the
actual delay between posting the scroll task and when it runs.

Enable layer scrolling to make this more efficient.

Design doc: go/bubble-launcher-scrolling
Bug: 1236207
Test: added to ash_unittests

Change-Id: I7674beb985c2a3f3f6b59b4ca4f6488e4458bcbb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3065272
Reviewed-by: Toni Baržić <tbarzic@chromium.org>
Commit-Queue: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/master@{#909017}