[CrOS Bluetooth] Add DeviceCache
This class is responsible for caching known paired Bluetooth devices so
that they can be returned by the CrosBluetoothConfig API. This class is
named DeviceCache instead of PairedDeviceCache because future CLs will
also utilize this class for the pairing flow, which deals with unpaired
devices.

This CL does not instantiate DeviceCache; a follow-up CL instantiates it
and uses it to return device metadata to clients.

Bug: 1010321
Change-Id: I9fdb72e776fa941533a5448721f4f01e3c693097
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3070213
Reviewed-by: Gordon Seto <gordonseto@google.com>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#909022}