Lookalikes: Add character swap heuristic
This CL adds a new heuristic for lookalike detection called character
swap. It detects neighboring characters that are swapped with each other
such as example[.]com vs exapmle[.]com.

These subtle changes in domain names may be hard to notice during casual
browsing. Early results indicate that this heuristic will only have a
small percentage of false positives.

The implementation in this CL is linear time, depending on the length of
the input strings. It works on the whole hostname instead of individual
domain labels.

This is the barebones implementation of the heuristic. It's currently
not used for lookalike detection. Follow up CLs will use the heuristic
to match navigated URLs against top domains or engaged sites.

Bug: 1109056
Change-Id: Ibcea92b89313d0e3311cb1158c0f477c4323ef65
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3069397
Commit-Queue: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Joe DeBlasio <jdeblasio@chromium.org>
Cr-Commit-Position: refs/heads/master@{#909044}