Add Shortcuts Widget Message strings.
This change adds Shortcuts Widget-specific Message strings to
announce that a specific functionality the user tries to interact with
may be unavailable on the device.

These messages are intended to be announced only when the
functionality has been disabled ahead of us detecting this. In most
other cases we will remove the relevant button from the Widget.

Mocks: http://shortn/_p89Zz1gmjZ

Bug: 1213541
Change-Id: I86997a0b8c63c3d9523e7f9b80ec4dd942d86dee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3075603
Commit-Queue: Tomasz Wiszkowski <ender@google.com>
Reviewed-by: Filip Gorski <fgorski@chromium.org>
Cr-Commit-Position: refs/heads/master@{#909069}