Set LPAC permissions on hard linked files on bots.
The luci infrastructure uses a cache for some swarming files and
hard links from the cache to the test directory.

This causes the script that sets LPAC ACLs to not successfully
set the right ACLs on these files.

This addition to the set_lpac_acls function  means that, when
running as part of a test, these files are identified and
the correct ACL is written to them.

BUG=1225029

Change-Id: I35d89c83e906f6cc880908ec43d8a470c4bb8ae0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3075257
Reviewed-by: Dirk Pranke <dpranke@google.com>
Commit-Queue: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#909114}