[CrOS WebAPKs] Add a policy to enable/disable WebAPKs
This adds a policy which allows admins to disable the Web APKs feature
to prevent Web App metadata from being sent to Google. The feature
defaults to enabled when no policy is set.

The policy is not yet enabled (it's `future_on` rather than
`supported_on`), I'll submit a CL to change this when we have full
launch approval.

Bug: 1232311
Change-Id: I9e141dfd5a22b21ff4076732bd9361c898b80df9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3066705
Commit-Queue: Tim Sergeant <tsergeant@chromium.org>
Reviewed-by: Melissa Zhang <melzhang@chromium.org>
Reviewed-by: Anqing Zhao <anqing@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#909132}