CCA: Better CSS extern types for closure compiler
The current closure compiler extern types for CSS Typed OM Level 1 has
some incorrect types according to spec [1]. Fix the types.

This also adds proper @property definitions for the custom properties.

This drops the error of `cca.py tsc` from 92 to 88 if we patched in
@types/w3c-css-typed-object-model-level-1. Unfortunately the error count
increases from 119 to 121 without the type definition because of the
missing types. All the CSSOM types for TypeScript will be addressed in
the later CL.

[1]: https://drafts.css-houdini.org/css-typed-om

Bug: b:172340451
Test: cca.py tsc
Test: closure compiler passes
Test: tast run <ip> camera.CCAUISmoke* camera.CCAUIStress*
Change-Id: Ib4a4b1c188b9c01c131b9cd4941134275a57a869
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3074700
Commit-Queue: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Shik Chen <shik@chromium.org>
Cr-Commit-Position: refs/heads/master@{#909235}