Android test runner: Add GTest option to skip upload.
When running the same multiple times for performance measurements,
the test APK and data upload steps can increase the total time for
measurement by a factor of 3 or more. Add an option to skip upload
and APK installation for use cases where we know the data on the
device will be unchanged.

Context: an ANGLE script that runs the same test suite multiple
times on an Android device, and restarts the test runner several
times to try and produce a measurement with lower inter-run
variance.

In local testing this sped up the total test time from ~200s to
~30s.

Bug: angleproject:6090
Change-Id: I7b4f8cce116cc0d3a5244315cc14029e7a17ab04
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3067515
Commit-Queue: Jamie Madill <jmadill@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/master@{#909284}