app_list: Close AppListBubble on click in bottom-left corner of screen
The shelf home button has a custom view targeter that extends its hit
test bounds to the corner of the screen. When deciding whether or not
to close the bubble we need to do a real hit test against the button,
not just check if the click point is inside its bounds.

Fixed: 1237264
Test: added to ash_unittests
Change-Id: Ida44d2f5272891c534da77b1324b40d0b9d8d97d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3077299
Commit-Queue: Toni Baržić <tbarzic@chromium.org>
Auto-Submit: James Cook <jamescook@chromium.org>
Reviewed-by: Toni Baržić <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#909348}