media/gpu/v4l2: add queue TRACE_COUNTER and use a small_map
While running the previous CL in Tast [2] I saw that the numbers
where very high (~300ms) and grew from the beginning of the test, so
I took a look at the size of the input and output (OUTPUT_ and
CAPTURE_queues, resp.) and they were indeed large (this might be
WAI). This CL adds a TRACE_COUNTER_ID2 for the size of those queues.

Since I touched abit V4L2Queue::queued_buffers_, I made it a small_map
(most of the time it only holds a bunch of element) and simplified a
few DCHECKs.

Bug: b:190615065
Change-Id: I070667f7d67723858103f20695298a8eb5c87097
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3072961
Commit-Queue: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Fritz Koenig <frkoenig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#909355}