Disable thread asserts for AwUncaughtExceptionTest.
Due to the complexities of the test, it triggers an thread assertion
during shutdown as it tries to reset the UI looper override.

As this test is explicitly testing uncaught exception handling,
disabling this thread assert does not fundamentally alter the
purpose of the testing code.

BUG=1236782

Change-Id: Ic5ea7dd31fcfa967e5c94e6fe828cce5a112ca06
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3078406
Reviewed-by: Bo <boliu@chromium.org>
Commit-Queue: Ted Choc <tedchoc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#909373}