Modify the full restore notification.
Change the full restore notification based on the latest ui decision:
https://docs.google.com/presentation/d/1gZbfpG_-4Sum7-ETMcV8-gOXxjd97mhacA__uiCjM88/edit#slide=id.ge7aea646b0_8_5

1. The Cancel button of the restore notification is changed from Cancel
to Settings. If the user chooses “Settings”, open the settings app, to
show the restore setting, and not close the notification.

The Cancel button of the crash notification is not changed, still the
Cancel button.

2. If user clicks the notification body, restore the last session, and
close the notification.

This change applied to both the restore notification and the crash
notification.

BUG=1146900

Change-Id: If139b991159a35d7ff7a686c2a460078524894e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3074640
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#909388}