Add learn more link to HFM interstitial
This updates the HTTPS-First Mode interstitial primary paragraph string
to include a "Learn more" link (matching how the SSL interstitial
strings are constructed), adds a CMD_OPEN_HELP_CENTER handler to
HttpsOnlyModeBlockingPage that opens the relevant help center link, and
a test that the learn more link works as expected.

This also adds an `IsShowingHttpsFirstModeInterstitial()` helper and
updates tests that were directly querying for the presence of the full
string (which is now broken due to how HTML in strings works).

Bug: 1218526
Change-Id: I20e9412495d94505a48c0cfac7b96d7892027776
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3076810
Reviewed-by: Carlos IL <carlosil@chromium.org>
Commit-Queue: Chris Thompson <cthomp@chromium.org>
Cr-Commit-Position: refs/heads/master@{#909392}