PagedViewStructure copy should not copy sanitize locks state
PagedViewStructure scoped sanitize locks are associated with the
originating structure, so the lock count should not be propagated to a
PagedViewStructure copy. This CL deletes the copy constructor, as
sanitize locks make the copy operation less intuitive (e.g. it's unclear
whether the copied structure should be sanitized or not). Instead, add a
method to load the view structure from another view structure associated
with the same apps grid. This way, code that copies the structure can
decide whether/when to sanitize the view structure (by setting needed
sanitize locks before loading the pages, if needed).

TEST=existing apps grid tests pass
(not really an issue in production at the moment, as we never copy view
structure while sanitize locks are active; an upcoming change may hit
this case, though)

BUG=1234002

Change-Id: I08421525bd12588919f87bd31540d870770fd724
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3076823
Commit-Queue: Toni Baržić <tbarzic@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/master@{#909407}