[SH iOS] Restructure click handlers
This CL continues restructuring work on the iOS Text Fragments
implementation. A new click handler is added, which is now responsible
for removing the marks in the page when the user clicks. This adds a
roundtrip from JS to Obj-C++ and back, but it will be necessary when
future patches on the native side need to hook into this process.

Bug: 1230574
Change-Id: I99b2f946bfd514b1acece2072993bd97ff1d8068
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3072040
Commit-Queue: Tommy Martino <tmartino@chromium.org>
Reviewed-by: Guillaume Jenkins <gujen@google.com>
Cr-Commit-Position: refs/heads/master@{#909409}