desk template: Get window's app_id from FullRestoreSaveHandler.
When opening some websites (e.g, Youtube) in a browser window,
ash::kAppIDKey associated with the browser window will change from
kChromeAppId to a different app id (in Youtube's case, it's
kYoutubeAppId). Thus when capturing a desk template, if we use
ash::kAppIDKey as the key to capture each window, we won't be able to
find its corresponding AppRestoreData and also won't be able to launch
it properly. In this case we change to get the window's full restore
app id from FullRestoreSaveHandler so that it's guaranteed that the
correct app id is returned.

Bug: 1234540
Change-Id: Ia6d7dbd903ba9d699dcc205f9b9b1c3bf01fe18d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3069751
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Nancy Wang <nancylingwang@chromium.org>
Commit-Queue: Xiaoqian Dai <xdai@chromium.org>
Cr-Commit-Position: refs/heads/master@{#909448}