Choose ETC1 vs ETC2 based on format returned by skia
Skia currently defaults to ETC2 format when passed ETC1 as both are
similar in terms of usage. This fails some checks and asserts as
some shared images expect ETC1 as there is no ETC2 in viz. To avoid
this, we check the returned format by skia and map the formats
accordingly.

Bug: 1236874
Change-Id: I71b39a2996b57786f92e3d63de0d60129a7b6107
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3076945
Commit-Queue: Saifuddin Hitawala <hitawala@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#909473}