[LayoutNG] Fragmentainer break tokens during OOF layout
The test case in Issue 1200328 was originally broken because we weren't
adding a break token to the last existing column when adding a new
column to the end for an OOF. That test case is no longer affected
by this problem, so this CL adds a number of unit tests to make sure
that this is working correctly to prevent this from breaking unknown or
future cases that may depend on this.

To fix this, logic was added to NGOutOfFlowLayoutPart::
LayoutOOFsInFragmentainer() to relayout the last existing fragmentainer
using the simplified OOF layout algorithm to create a break token if
we are adding a new fragmentatiner, and the last fragmentainer had no
break token previously.

In addition to adding a fragmentainer break token in this case, this
CL also ensures that any break tokens created for new fragmentainers as
a result of OOF layout has the correct consumed block size and flags
set. To fix this, a new method was added, called
FinishFragmentationForFragmentainer(), that applies the info needed
to create an accurate break token to a given fragmentainer. This is
now called before completing layout in NGSimplifiedOOFLayoutAlgorithm.

Additionally, if a column was updated to hold an OOF, and it
had broken previously due to a column spanner, we would fail to
reflect this in the new break token that was created. To fix this,
add an additional flag to the NGBoxFragmentBuilder to indicate if
a column spanner was found, even if we don't have reference to that
spanner when running the simplified layout.

Bug: 1200328
Change-Id: I9de620d4833991e06beb1447d9d5fe209decabf4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3076808
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#909501}