[Chromecast] Create ExternalConnector for media task runner
This fixes a check failure when ExternalConnector is used on the media
task runner. We clone the main ExternalConnector early in the browser
lifecycle. When CastRenderer makes an interface request, the cloned
connector is bound to the media task runner.

Bug: b/195704841
Test: Launch media on cast_shell
Change-Id: If0cb53d82f8384bd174148574110937d6823ef9a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3078899
Auto-Submit: Sean Topping <seantopping@chromium.org>
Reviewed-by: Kenneth MacKay <kmackay@chromium.org>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Commit-Queue: Greg Kerr <kerrnel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#909531}