Android: fix existing @CommandLineFlags usage
This adds the CommandLineFlags TestRule to several classes to ensure
the CommandLine switches are actually changed by the annotation. This
switches to BaseRobolectricTestRunner because this is required in order
to use CommandLineFlags.

This CL fixes the tests in /chrome/browser/download.

This CL was uploaded by git cl split.

R=dtrainor@chromium.org

Bug: 1236779
Test: run_chrome_junit_tests
Change-Id: I2574abc27aa142f2f75009cc27e1802086ce5638
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3072963
Auto-Submit: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Commit-Queue: Nate Fischer <ntfschr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#909541}