[CapturePicker] Update text and wrapping for Capture Picker
After consulting with UX, it was decided that the string shown when
capture choices are limited by enterprise policy should indicate that
some choices may have been hidden. The goal of this is to prevent
confusion when tabs may be hidden from the tab list portion of the
picker. While making this update, I added some additional padding to the
string and discovered that it was being cut off, so this also updates
the string to properly wrap when it overflows.

Change-Id: Ice30ac5a5e06c09c1b4f0eab49345bc3626f8f68
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3078480
Auto-Submit: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/master@{#909559}