indexeddb: bump histograms until next year
IDB has a lot of histograms, but they are often things that are useful
to check periodically when considering bugs.  I think it's worth
continuing to track most of these metrics.

As per the histogram guidance, "1 year" is a reasonable timeframe to
bump histograms that are indefinitely useful.

Bug: 975159, 1081299, 1081302, 1086222
Change-Id: I571221669c82875266d185621dc0d5a17f286975
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3076088
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Commit-Queue: Stephen Chenney <schenney@chromium.org>
Auto-Submit: enne <enne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#909571}