gpu: Keep DC layer overlay access in SkiaOutputDeviceGL
Make SkiaOutputDeviceGL own the ScopedOverlayAccess and keep it until
SwapBuffers completes.  This has the same effect as the previous code
which passed along the scoped access in the DC layer params release
callback so it removes the need for that callback.  This approach is
better since the GLSurface is a ref counted object which could be kept
alive in general after the SkiaOutputDeviceGL is gone, which would also
make the scoped access outlive the shared image representation causing
the CHECK failure in the bug.

Bug: 1227771
Change-Id: I43bdd1f92b5cdfae57ec9aecc0cdfdfd75f0beb2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3063792
Auto-Submit: Sunny Sachanandani <sunnyps@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Sunny Sachanandani <sunnyps@chromium.org>
Cr-Commit-Position: refs/heads/master@{#909595}