[Keyboard Tooltip] Always UpdateTooltipFromKeyboard if focus set from keypress
This CL fixes a bug where a keyboard-triggered tooltip isn't displayed
when it should. Consider the following HTML:

<button title="tooltip">hello</button>
<button>world</button> <!-- no title -->

If we were to tab to the first button, the tooltip would show up.
However, after tabbing to the second button and shift + tab back to the
first button, the tooltip wouldn't show up.

This happens because the TooltipController doesn't get notified that the
focus moved to another control that was devoided of tooltip text. It
thinks that we are still on that same button, thus that we shouldn't
show the tooltip again.

This CL always calls UpdateTooltipFromKeyboard regardless of whether it
has tooltip text or not.

Bug: 829352
Change-Id: I4786aade06a08633c4031222291b115b01ab9fae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3079099
Commit-Queue: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Reviewed-by: Daniel Libby <dlibby@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#909612}