Ensure Price Drop URL is in sync with Tab URL
It is proving untenable to keep the price drop URL in sync with the Tab
URL using TabObserver as it can behave unexpectedly in the wild. For
example, an iframe counts as a new navigation and it is difficult to
guarantee we filter these out. A better approach is to store the Tab URL
at the time of price drop acquisition and compare that with the Tab URL
when getPriceDrop() is called as well as when we are determining if we
should re-fetch the ShoppingPersistedTabData.

Bug: 1237304
Change-Id: I31b4873bdfef638be707870b1773861fa33bb55f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3076969
Reviewed-by: David Trainor <dtrainor@chromium.org>
Commit-Queue: David Maunder <davidjm@chromium.org>
Cr-Commit-Position: refs/heads/master@{#909653}