Use InstanceKey in more places in AppServiceInstanceRegistryHelper
aura::Window* will not be available in Ash for web apps that run in
tabs after Lacros is enabled. Any code that uses an aura::Window* to
refer to an instance should instead use an InstanceKey unless it is
guaranteed to only be referring to top level aura::Window*s (which
will still be available in Ash with Lacros).

This converts CalculateVisibilityState, CalculateActivatedState,
GetAppId and GetState member functions to take an InstanceKey rather
than an aura::Window*.

BUG=1235206

Change-Id: I7ac9dc5897a7fba48fdc555a50f30882731d23c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3071502
Reviewed-by: Nancy Wang <nancylingwang@chromium.org>
Commit-Queue: Raymes Khoury <raymes@chromium.org>
Cr-Commit-Position: refs/heads/master@{#909686}