[Lacros] Add WebAppSyncBridge DCHECKs
We DCHECK that non-system web apps are not added to the
registrar if web apps are not user installable.

Most specifically, when the Lacros flags are enabled,
non-system web apps must not be added to the registrar
in Ash.

Bug: 1207199
Change-Id: I3a951f708ed3ac2adc33b657fdc57da4bbf90cfe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3076982
Auto-Submit: Eric Willigers <ericwilligers@chromium.org>
Commit-Queue: Dominic Schulz <dominicschulz@google.com>
Reviewed-by: Dominic Schulz <dominicschulz@google.com>
Cr-Commit-Position: refs/heads/master@{#909693}